==> 3.BioData_pipe1_MatrixFormation.ipynb <==
def GenerateMatrix (Moore_VCF, Clone_no, **kwargs):
    import pandas as pd

    SigProfMatrix = pd.DataFrame (index = range ( len (Moore_VCF)),  columns = ["Project", "Sample", "ID", "Genome", "mut_type", "chrom",	"pos_start","pos_end",	"ref",	"alt"	,"Type"]) 
    SigProfMatrix["Project"] = "BioData"
    SigProfMatrix["Sample"] = str(Clone_no)
    SigProfMatrix["ID"] = Moore_VCF["pos"]
    SigProfMatrix["Genome"] = "GRCh37"
    SigProfMatrix["mut_type"] = "SNP"
    SigProfMatrix["chrom"] = Moore_VCF["chr"]
    SigProfMatrix["pos_start"] = Moore_VCF["start"]
    SigProfMatrix["pos_end"] = Moore_VCF["end"]
    SigProfMatrix["ref"] = Moore_VCF["ref"]
    SigProfMatrix["alt"] = Moore_VCF["alt"]
    SigProfMatrix["Type"] = "SOMATIC"

    pd.DataFrame(SigProfMatrix).to_csv( kwargs["OUTPUT_PATH"], index=False, header = True,  sep="\t")


def extract (Moore_VCF, **kwargs):
    import numpy as np

    # Clone number 대로  mutation_id를 뽑아주기
    with open( kwargs ["DEICISION_MEMBERSHIP_PATH"], "r") as input_file:
        membership = np.array( [int(line.strip()) for line in input_file] )
        npvaf = pd.read_csv ( kwargs ["NPVAF_PATH"], sep = "\t")
        for Clone_no in sorted(np.unique ( membership) ) :
            npvaf_Clone_no =  (npvaf.iloc [np.where(membership == Clone_no)[0] ])
            pos_Clone_no =  list ( npvaf_Clone_no.iloc[:, 0] )
            print (pos_Clone_no)

            Moore_VCF_Clone_no = Moore_VCF[Moore_VCF['pos'].isin( pos_Clone_no ) == True]
            
            kwargs["OUTPUT_PATH"] = kwargs["OUTPUT_DIR"] + "/clone{}.txt"
            GenerateMatrix (Moore_VCF_Clone_no, Clone_no, **kwargs)

    # whole
    kwargs["OUTPUT_PATH"] = kwargs["OUTPUT_DIR"] + "/whole.txt"
    GenerateMatrix (Moore_VCF, "whole", **kwargs)


if __name__ == "__main__":
    from SigProfilerMatrixGenerator.scripts import SigProfilerMatrixGeneratorFunc as matGen
    import pandas as pd
    import argparse

    Moore_VCF = pd.read_csv("/data/project/Alzheimer/CLEMENT/resource/paper/whole_info.txt", sep = "\t")

    # column을 소문자로 바꾸기
    Moore_VCF.rename(columns=lambda x: x.lower(), inplace=True)

    # lexicographically 정렬하기
    custom_order = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "X", "Y"]
    Moore_VCF['chr_categorical'] = pd.Categorical(Moore_VCF['chr'], categories=custom_order, ordered=True)
    Moore_VCF = Moore_VCF.sort_values(by= ['chr_categorical', 'start']).reset_index(drop=True).drop ("chr_categorical", axis = 1)

    # 앞에 "chr" 붙여주기
    Moore_VCF["pos"]  = "chr" + Moore_VCF["chr"] + "_" + Moore_VCF["start"].astype("str")


    kwargs = {}
    kwargs ["DEICISION_MEMBERSHIP_PATH"] = "/data/project/Alzheimer/CLEMENT/03.combinedoutput/3.BioData/Moore_1D/adrenal_gland_zona_glomerulosa/PD28690-L1/CLEMENT_hard_1st.membership.txt"
    kwargs ["NPVAF_PATH"] = "/data/project/Alzheimer/CLEMENT/02.npvaf/3.BioData/Moore_1D/adrenal_gland_zona_glomerulosa/PD28690-L1/npvaf.txt"


    extract (Moore_VCF, **kwargs)

from SigProfilerMatrixGenerator.scripts import SigProfilerMatrixGeneratorFunc as matGen
import pandas as pd

Moore_VCF = pd.read_csv("/data/project/Alzheimer/CLEMENT/resource/paper/whole_info.txt", sep = "\t")

# column을 소문자로 바꾸기
Moore_VCF.rename(columns=lambda x: x.lower(), inplace=True)

# lexicographically 정렬하기
custom_order = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "X", "Y"]
Moore_VCF['chr_categorical'] = pd.Categorical(Moore_VCF['chr'], categories=custom_order, ordered=True)
Moore_VCF = Moore_VCF.sort_values(by= ['chr_categorical', 'start']).reset_index(drop=True).drop ("chr_categorical", axis = 1)

# 앞에 "chr" 붙여주기
Moore_VCF["pos"]  = "chr" + Moore_VCF["chr"] + "_" + Moore_VCF["start"].astype("str")

# Filter rows where 'pos' is in the list 'a'
list_of_interest = ["chr1_115672", "chr1_525383", "chr1_540719"]
Moore_VCF = Moore_VCF[Moore_VCF['pos'].isin( list_of_interest ) == True]
Moore_VCF

SigProfMatrix = pd.DataFrame (index = range ( len (Moore_VCF)),  columns = ["Project", "Sample", "ID", "Genome", "mut_type", "chrom",	"pos_start","pos_end",	"ref",	"alt"	,"Type"]) 
SigProfMatrix["Project"] = "BioData"
SigProfMatrix["Sample"] = "clone1"
SigProfMatrix["ID"] = Moore_VCF["pos"]
SigProfMatrix["Genome"] = "GRCh37"
SigProfMatrix["mut_type"] = "SNP"
SigProfMatrix["chrom"] = Moore_VCF["chr"]
SigProfMatrix["pos_start"] = Moore_VCF["start"]
SigProfMatrix["pos_end"] = Moore_VCF["end"]
SigProfMatrix["ref"] = Moore_VCF["ref"]
SigProfMatrix["alt"] = Moore_VCF["alt"]
SigProfMatrix["Type"] = "SOMATIC"

SigProfMatrix
